+++ usage.py
@@ -14,7 +14,7 @@ from dash.dependencies import Output
 app = dash.Dash(__name__)
 
 UPLOAD_FOLDER_ROOT = (
-    r"C:\tmp\Uploads" if sys.platform.startswith("win") else "/tmp/Uploads"
+    R"C:\tmp\Uploads" if sys.platform.startswith("win") else "/tmp/Uploads"
 )
 du.configure_upload(app, UPLOAD_FOLDER_ROOT)
 
@@ -69,7 +69,7 @@ app.layout = get_app_layout
     id="dash-uploader",
 )
 def callback_on_completion(status: du.UploadStatus):
-    return html.Ul([html.Li(str(x)) for x in status.uploaded_files])
+    return html.Ul([html.Li(str(x)) for x in status["uploaded_files"]])
 
 
 if __name__ == "__main__":